Compute targets of a fresh LetVar.
